fix: ajuste de componentes en el modulo solicitudesgiros #8
