--- src/app/pages/consultas-solicitudesgiros/components/layout/layout.component.html
@@ -1,20 +1,0 @@
-<div class="container">
-    <div class="row">
-        <div class="col py-2"></div>
-    </div>
-    <div class="row">
-        <div class="col-12 col-sm-12 col-md-12 pr-0">
-            <div class="container-fluid">
-                <div class="row">
-                    <div class="col">
-                        <div class="card">
-                            <div class="card-body">
-                                <ngx-table-listasolicitudes></ngx-table-listasolicitudes>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
