--- src/app/pages/consultas-solicitudesgiros/solicitudesgiros.module.ts
@@ -1,29 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { SolicitudesgirosRoutingModule } from './solicitudesgiros-routing.module';
-import { LayoutComponent } from './components/layout/layout.component';
-import { StoreModule } from '@ngrx/store';
-import * as fromSolicitudesgiros from './reducers/solicitudesgiros.reducer';
-import { EffectsModule } from '@ngrx/effects';
-import { SolicitudesgirosEffects } from './effects/solicitudesgiros.effects';
-import { TableListasolicitudesComponent } from './components/table-listasolicitudes/table-listasolicitudes.component';
-import { SharedModule } from '../../shared/shared.module';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
-
-@NgModule({
-  declarations: [LayoutComponent,
-    TableListasolicitudesComponent,
-  ],
-  imports: [
-    CommonModule,
-    SolicitudesgirosRoutingModule,
-    SharedModule,
-    FormsModule,
-    ReactiveFormsModule,
-    StoreModule.forFeature(fromSolicitudesgiros.solicitudesgirosFeatureKey, fromSolicitudesgiros.reducer),
-    EffectsModule.forFeature([SolicitudesgirosEffects])
-  ]
-})
-export class SolicitudesgirosModule { }
