--- src/app/pages/crear-solicitudgiro/components/layout/layout.component.scss
@@ -1,7 +1,0 @@
-.mat-stepper-horizontal {
-    margin-top: 8px;
-}
-
-.mat-form-field {
-    margin-top: 16px;
-}
