--- src/app/pages/crear-solicitudgiro/components/show-resumensolicitudgiro/show-resumensolicitudgiro.component.html
@@ -1,3 +1,0 @@
-<form [formGroup]="resumenSolicitudGroup">
-    <p>show-resumensolicitudgiro works!</p>
-</form>
