+++ src/app/pages/solicitudesgiros/components/show-resumensolicitudgiro/show-resumensolicitudgiro.component.html
@@ -1,0 +1,27 @@
+<form [formGroup]="resumenSolicitudGroup">
+    <p>show-resumensolicitudgiro works!</p>
+    <!-- Grupo: Información de Solicitud -->
+    <div class="row">
+        <div class="col-md-12">
+            <mat-card-header class="py-1">
+                <mat-card-subtitle>
+                    <h6 class="card-title col-md-4 bg-primary text-white text-center py-2 px-2 rounded">Información de la Solicitud</h6>
+                </mat-card-subtitle>
+            </mat-card-header>
+        </div>
+    </div>
+    <div class="row px-5">
+        <div class="col-md-12 py-2">
+            <mat-form-field appearance="outline" [style.width.%]="100" required>
+                <mat-label>Concepto</mat-label>
+                <mat-select disableRipple matInput [ngModel] formControlName="concepto" required>
+                    <mat-option value="1">Option 1: Lorem ipsum dolor sit, amet consectetur adipisicing elit</mat-option>
+                    <mat-option value="2">Option 2: Lorem ipsum dolor sit, amet consectetur adipisicing elit</mat-option>
+                    <mat-option value="3">Option 3: Lorem ipsum dolor sit, amet consectetur adipisicing elit</mat-option>
+                </mat-select>
+                <mat-hint>Elegir concepto de solicitud</mat-hint>
+            </mat-form-field>
+        </div>
+    </div>
+    <!-- Fin Campo: Elegir Concepto -->
+</form>
