--- src/app/shared/components/datos-generales-iniciales/datos-generales-iniciales.component.html
@@ -1,29 +1,0 @@
-<div class="container-fluid">
-    <div class="row m-0 p-0">
-        <div class="col m-0 p-0">
-            <div class="card">
-                <div class="card-body">
-                    <h5 *ngIf="Title" class="card-title text-center text-primary">{{Title}}</h5>
-                    <!-- <div class="container">
-                        <div class="row">
-                            <div class="col-6 d-flex justify-content-center">
-                                <div class="form-group">
-                                    <label>Centro Gestor:</label>
-                                    <input type="text" class="form-control py-2" id="buscador-tabla"
-                                        aria-describedby="search-table" placeholder="">
-                                </div>
-                            </div>
-                            <div class="col-6 d-flex justify-content-center">
-                                <div class="form-group">
-                                    <label>Area Funcional:</label>
-                                    <input type="text" class="form-control py-2" id="buscador-tabla"
-                                        aria-describedby="search-table" placeholder="">
-                                </div>
-                            </div>
-                        </div>
-                    </div> -->
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
